Make all sound levels float64
A union of integer and float level arrays in SoundPulses was an
unnecessary complication. Levels are not on the exact emulated time
axis; they are bound for float samples anyway, and 0 and 1 are
exact in float64, so binary emitters lose nothing.

Verified with a full elven_war.rzx playback.

+++ tests/test_sound.py
@@ -38,7 +38,7 @@ class _RecordingSound(SoundDevice):
 
 # A chunk holding a single level over its whole span.
 def _level_chunk(rate: int, level: int, num_ticks: int) -> SoundPulses:
-    levels = numpy.array([level], dtype=numpy.uint32)
+    levels = numpy.array([level], dtype=numpy.float64)
     ticks = numpy.array([0], dtype=numpy.uint32)
     return SoundPulses(rate, levels, ticks, num_ticks=num_ticks)
